SecurityToolSuite/requirements.txt
cryptography>=3.4.7
requests>=2.26.0
python-nmap>=0.7.1
tkinter
